Revert "fix: Prevent chart popovers from overflowing horizontally (#1804)"
This reverts commit 137bd15ed8212f54df8f266e6e406866e513bf25.
