+++ src/internal/components/chart-popover/index.tsx
@@ -98,9 +98,6 @@ function ChartPopover(
     };
   }, [container, onDismiss]);
 
-  // In chart popovers, dismiss button is present when they are pinned, so both values are equivalent.
-  const isPinned = dismissButton;
-
   return (
     <div
       {...baseProps}
@@ -126,8 +123,6 @@ function ChartPopover(
           </div>
         )}
         keepPosition={true}
-        allowVerticalOverflow={true}
-        allowScrollToFit={isPinned}
       >
         <div className={styles['hover-area']}>
           <PopoverBody
