--- src/internal/utils/__tests__/scrollable-containers.test.ts
@@ -1,29 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-import { scrollRectangleIntoView } from '../scrollable-containers';
-
-const originalScrollBy = window.scrollBy;
-
-beforeEach(() => {
-  window.scrollBy = jest.fn();
-  window.innerHeight = 500;
-});
-afterEach(() => {
-  window.scrollBy = originalScrollBy;
-});
-
-describe('scrollRectangleIntoView', () => {
-  test("scrolls up until the rectangle's top position fits the viewport top, if the rectangle's top position was further up", () => {
-    scrollRectangleIntoView({ top: -50, height: 100, width: 100, left: 0 });
-    expect(window.scrollBy).toHaveBeenCalledWith(0, -50);
-  });
-  test("scrolls down until the rectangle's bottom position fits the viewport bottom, if the rectangle's bottom position was further down", () => {
-    scrollRectangleIntoView({ top: 300, height: 300, width: 100, left: 0 });
-    expect(window.scrollBy).toHaveBeenCalledWith(0, 100);
-  });
-  test("scrolls down only until the rectangle's top position fits the viewport top, if the rectangle's bottom position was further down but it is taller than the viewport", () => {
-    scrollRectangleIntoView({ top: 300, height: 600, width: 100, left: 0 });
-    expect(window.scrollBy).toHaveBeenCalledWith(0, 300);
-  });
-});
