+++ src/internal/utils/scrollable-containers.ts
@@ -1,6 +1,6 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-export interface BoundingBox {
+export interface Dimensions {
   height: number;
   width: number;
   top: number;
@@ -27,7 +27,7 @@ export const getOverflowParentDimensions = ({
   excludeClosestParent: boolean;
   expandToViewport: boolean;
   canExpandOutsideViewport: boolean;
-}): BoundingBox[] => {
+}): Dimensions[] => {
   const parents = expandToViewport
     ? []
     : getOverflowParents(element).map(el => {
@@ -83,29 +83,4 @@ export function scrollElementIntoView(
   options: ScrollIntoViewOptions = { block: 'nearest', inline: 'nearest' }
 ) {
   element?.scrollIntoView?.(options);
-}
-
-export function calculateScroll({ top, height }: BoundingBox) {
-  if (top < 0) {
-    return top;
-  } else if (top + height > window.innerHeight) {
-    if (height > window.innerHeight) {
-      return top;
-    } else {
-      return top + height - window.innerHeight;
-    }
-  }
-  return 0;
-}
-
-/**
- * For elements with fixed position, the browser's native scrollIntoView API doesn't work,
- * so we need to manually scroll to the element's position.
- * Supports only vertical scrolling.
- */
-export function scrollRectangleIntoView(box: BoundingBox) {
-  const scrollAmount = calculateScroll(box);
-  if (scrollAmount) {
-    window.scrollBy(0, scrollAmount);
-  }
 }
