+++ src/popover/utils/positions.ts
@@ -1,18 +1,18 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-import { PopoverProps, InternalPosition, BoundingBox, Dimensions } from '../interfaces';
+import { PopoverProps, InternalPosition, BoundingOffset, BoundingBox } from '../interfaces';
 
 // A structure describing how the popover should be positioned
 interface CalculatedPosition {
   scrollable?: boolean;
   internalPosition: InternalPosition;
-  rect: BoundingBox;
+  boundingOffset: BoundingOffset;
 }
 
 interface ElementGroup {
-  body: Dimensions;
-  trigger: BoundingBox;
-  arrow: Dimensions;
+  body: BoundingBox;
+  trigger: BoundingOffset;
+  arrow: BoundingBox;
 }
 
 const ARROW_OFFSET = 12;
@@ -68,7 +68,7 @@ export const PRIORITY_MAPPING: Record<PopoverProps.Position, InternalPosition[]>
   ],
 };
 
-const RECTANGLE_CALCULATIONS: Record<InternalPosition, (r: ElementGroup) => BoundingBox> = {
+const RECTANGLE_CALCULATIONS: Record<InternalPosition, (r: ElementGroup) => BoundingOffset> = {
   'top-center': ({ body, trigger, arrow }) => {
     return {
       top: trigger.top - body.height - arrow.height,
@@ -151,7 +151,19 @@ const RECTANGLE_CALCULATIONS: Record<InternalPosition, (r: ElementGroup) => Boun
   },
 };
 
-function fitIntoContainer(inner: BoundingBox, outer: BoundingBox): BoundingBox {
+/**
+ * Returns whether one rectangle fits in another.
+ */
+function canRectFit(inner: BoundingOffset, outer: BoundingOffset): boolean {
+  return (
+    inner.left >= outer.left &&
+    inner.top >= outer.top &&
+    inner.left + inner.width <= outer.left + outer.width &&
+    inner.top + inner.height <= outer.top + outer.height
+  );
+}
+
+function fitIntoContainer(inner: BoundingOffset, outer: BoundingOffset): BoundingOffset {
   let { left, width, top, height } = inner;
 
   // Adjust left boundary.
@@ -176,43 +188,32 @@ function fitIntoContainer(inner: BoundingBox, outer: BoundingBox): BoundingBox {
   return { left, width, top, height };
 }
 
-function getTallestRect(rect1: BoundingBox, rect2: BoundingBox): BoundingBox {
-  return rect1.height >= rect2.height ? rect1 : rect2;
-}
-
-function getIntersection(rectangles: BoundingBox[]): BoundingBox | null {
-  let boundingBox: BoundingBox | null = null;
+/**
+ * Returns the area of the intersection of passed in rectangles or a null, if there is no intersection
+ */
+export function intersectRectangles(rectangles: BoundingOffset[]): number | null {
+  let boundingOffset: BoundingOffset | null = null;
   for (const currentRect of rectangles) {
-    if (!boundingBox) {
-      boundingBox = currentRect;
+    if (!boundingOffset) {
+      boundingOffset = currentRect;
       continue;
     }
-    const left = Math.max(boundingBox.left, currentRect.left);
-    const top = Math.max(boundingBox.top, currentRect.top);
-    const right = Math.min(boundingBox.left + boundingBox.width, currentRect.left + currentRect.width);
-    const bottom = Math.min(boundingBox.top + boundingBox.height, currentRect.top + currentRect.height);
+    const left = Math.max(boundingOffset.left, currentRect.left);
+    const top = Math.max(boundingOffset.top, currentRect.top);
+    const right = Math.min(boundingOffset.left + boundingOffset.width, currentRect.left + currentRect.width);
+    const bottom = Math.min(boundingOffset.top + boundingOffset.height, currentRect.top + currentRect.height);
     if (right < left || bottom < top) {
       return null;
     }
-    boundingBox = {
+    boundingOffset = {
       left,
       top,
       width: right - left,
       height: bottom - top,
     };
   }
-  return boundingBox;
+  return boundingOffset && boundingOffset.height * boundingOffset.width;
 }
-
-/**
- * Returns the area of the intersection of passed in rectangles or a null, if there is no intersection
- */
-export function intersectRectangles(rectangles: BoundingBox[]): number | null {
-  const boundingBox: BoundingBox | null = getIntersection(rectangles);
-  return boundingBox && boundingBox.height * boundingBox.width;
-}
-
-type CandidatePosition = CalculatedPosition & { visibleArea: BoundingBox | null };
 
 /**
  * A functions that returns the correct popover position based on screen dimensions.
@@ -227,86 +228,55 @@ export function calculatePosition({
   viewport,
   // the popover is only bound by the viewport if it is rendered in a portal
   renderWithPortal,
-  allowVerticalOverflow,
 }: {
   preferredPosition: PopoverProps.Position;
   fixedInternalPosition?: InternalPosition;
-  trigger: BoundingBox;
-  arrow: Dimensions;
-  body: Dimensions;
-  container: BoundingBox;
-  viewport: BoundingBox;
+  trigger: BoundingOffset;
+  arrow: BoundingBox;
+  body: BoundingBox;
+  container: BoundingOffset;
+  viewport: BoundingOffset;
   // the popover is only bound by the viewport if it is rendered in a portal
   renderWithPortal?: boolean;
-  allowVerticalOverflow?: boolean;
 }): CalculatedPosition {
-  let bestOption: CandidatePosition | null = null;
+  let bestPositionOutsideViewport: CalculatedPosition | null = null;
+  let largestArea = 0;
 
   // If a fixed internal position is passed, only consider this one.
   const preferredInternalPositions = fixedInternalPosition
     ? [fixedInternalPosition]
     : PRIORITY_MAPPING[preferredPosition];
 
-  // Attempt to position the popover based on the priority list for this position.
+  // Attempt to position the popover based on the priority list for this position,
+  // trying to fit it inside the container and inside the viewport.
   for (const internalPosition of preferredInternalPositions) {
-    const rect = RECTANGLE_CALCULATIONS[internalPosition]({ body, trigger, arrow });
-    const visibleArea = renderWithPortal
-      ? getIntersection([rect, viewport])
-      : getIntersection([rect, viewport, container]);
-
-    const fitsWithoutOverflow = visibleArea && visibleArea.width === body.width && visibleArea.height === body.height;
-
-    if (fitsWithoutOverflow) {
-      return { internalPosition, rect };
+    const boundingOffset = RECTANGLE_CALCULATIONS[internalPosition]({ body, trigger, arrow });
+    const fitsInContainer = renderWithPortal || canRectFit(boundingOffset, container);
+    const fitsInViewport = canRectFit(boundingOffset, viewport);
+    if (fitsInContainer && fitsInViewport) {
+      return { internalPosition, boundingOffset };
     }
-
-    const newOption = { rect, internalPosition, visibleArea };
-    bestOption = getBestOption(newOption, bestOption);
+    const boundingRectangles = [boundingOffset, viewport];
+    if (!renderWithPortal) {
+      boundingRectangles.push(container);
+    }
+    const availableArea = intersectRectangles(boundingRectangles);
+    if (availableArea && availableArea > largestArea) {
+      bestPositionOutsideViewport = { internalPosition, boundingOffset };
+      largestArea = availableArea;
+    }
   }
 
   // Use best possible placement.
-  const internalPosition = bestOption?.internalPosition || 'right-top';
+  const internalPosition = bestPositionOutsideViewport?.internalPosition || 'right-top';
   // Get default rect for that placement.
-  const rect = RECTANGLE_CALCULATIONS[internalPosition]({ body, trigger, arrow });
+  const defaultOffset = RECTANGLE_CALCULATIONS[internalPosition]({ body, trigger, arrow });
+  // Get largest possible rect that fits into viewport.
+  const optimisedOffset = fitIntoContainer(defaultOffset, viewport);
+  // If largest possible rect is smaller than original - set body scroll.
+  const scrollable = optimisedOffset.height < defaultOffset.height;
 
-  // Get largest possible rect that fits into the viewport or container.
-
-  // We allow the popover to overflow the viewport if allowVerticalOverflow is true _and_ the popover will be anchored to the top or the bottom.
-  // If it is anchored to the right or left, we consider that it should have enough vertical space so that applying scroll to it is a better option.
-  const tallestBoundingContainer = getTallestRect(viewport, container);
-  const boundingContainer =
-    allowVerticalOverflow && isTopOrBottom(internalPosition)
-      ? {
-          top: tallestBoundingContainer.top,
-          height: tallestBoundingContainer.height,
-          left: viewport.left,
-          width: viewport.width,
-        }
-      : viewport;
-
-  const optimizedRect = fitIntoContainer(rect, boundingContainer);
-
-  // If largest possible rect is shorter than original - set body scroll.
-  const scrollable = optimizedRect.height < rect.height;
-
-  return { internalPosition, rect: optimizedRect, scrollable };
-}
-
-function getBestOption(option1: CandidatePosition, option2: CandidatePosition | null) {
-  // Within calculatePosition, the only case where option2 will not be defined will be in the first call.
-  // The only case where the visibleArea of an option will be null is when the popover is totally outside of the viewport.
-  if (!option2?.visibleArea) {
-    return option1;
-  }
-  if (!option1.visibleArea) {
-    return option2;
-  }
-  // Only if none of the two options overflows horizontally, choose the best based on the visible height.
-  if (option1.visibleArea.width === option2.visibleArea.width) {
-    return option1.visibleArea.height > option2.visibleArea.height ? option1 : option2;
-  }
-  // Otherwise, choose the option that is less cut off horizontally.
-  return option1.visibleArea.width > option2.visibleArea.width ? option1 : option2;
+  return { internalPosition, boundingOffset: optimisedOffset, scrollable };
 }
 
 export function getOffsetDimensions(element: HTMLElement) {
@@ -319,8 +289,4 @@ export function getDimensions(element: HTMLElement) {
     width: parseFloat(computedStyle.width),
     height: parseFloat(computedStyle.height),
   };
-}
-
-function isTopOrBottom(internalPosition: InternalPosition) {
-  return ['top', 'bottom'].includes(internalPosition.split('-')[0]);
 }
